+++ R/RcppExports.R
@@ -17,8 +17,8 @@ pagfl_routine <- function(y, X, method, Z, i_index, t_index, N, bias_correc, lam
     .Call(`_PAGFL_pagfl_routine`, y, X, method, Z, i_index, t_index, N, bias_correc, lambda_vec, kappa, min_group_frac, max_iter, tol_convergence, tol_group, varrho, rho, parallel, verbose)
 }
 
-tv_pagfl_routine <- function(y, X, X_const, d, M, i_index, t_index, N, p_const, lambda_vec, kappa, min_group_frac, max_iter, tol_convergence, tol_group, varrho, rho, parallel, verbose) {
-    .Call(`_PAGFL_tv_pagfl_routine`, y, X, X_const, d, M, i_index, t_index, N, p_const, lambda_vec, kappa, min_group_frac, max_iter, tol_convergence, tol_group, varrho, rho, parallel, verbose)
+fuse_time_routine <- function(y, X, X_const, d, M, i_index, t_index, N, p_const, lambda_vec, kappa, min_group_frac, max_iter, tol_convergence, tol_group, varrho, rho, parallel, verbose) {
+    .Call(`_PAGFL_fuse_time_routine`, y, X, X_const, d, M, i_index, t_index, N, p_const, lambda_vec, kappa, min_group_frac, max_iter, tol_convergence, tol_group, varrho, rho, parallel, verbose)
 }
 
 getTVAlpha <- function(xi, K_hat, p, n_periods, B) {