+++ tests/testthat/fixtures/test_helper.R
@@ -39,18 +39,18 @@ check_pagfl_output <- function(estim, X, i_index = NULL, t_index = NULL, oracle 
   }
 }
 
-check_tv_pagfl <- function(estim, groups_0, alpha_0) {
+check_fuse_time <- function(estim, groups_0, alpha_0) {
   expect_equal(estim$groups$groups, groups_0, ignore_attr = TRUE)
   expect_equal(estim$groups$n_groups, max(groups_0), ignore_attr = TRUE)
   expect_equal(estim$coefficients$tv[20:80, 2, ], alpha_0[20:80, 2, ], ignore_attr = TRUE, tolerance = 0.1)
   expect_equal(round(estim$IC$IC, 6), 0.313991)
   expect_equal(round(estim$IC$msr, 6), 1.053228)
-  resid_0 <- readRDS(test_path("fixtures", "tv_pagfl_resid.rds"))
+  resid_0 <- readRDS(test_path("fixtures", "fuse_time_resid.rds"))
   expect_equal(estim$residuals, resid_0)
 }
 
 
-check_tv_pagfl_output <- function(estim, X, i_index = NULL, t_index = NULL, oracle = FALSE) {
+check_fuse_time_output <- function(estim, X, i_index = NULL, t_index = NULL, oracle = FALSE) {
   alpha_hat <- estim$coefficients$tv
   expect_equal(colnames(alpha_hat)[-1], colnames(X))
   expect_equal(dimnames(alpha_hat)[[3]], paste("Group", 1:estim$groups$n_groups))
@@ -89,7 +89,7 @@ check_pagfl_sim <- function(sim, N, n_periods, p, q = NULL, alpha_0 = NULL, K) {
   }
 }
 
-check_tv_pagfl_sim <- function(sim, N, n_periods, p, K) {
+check_fuse_time_sim <- function(sim, N, n_periods, p, K) {
   expect_length(sim$y, N * n_periods)
   expect_equal(dim(sim$X), c(N * n_periods, p))
   expect_equal(max(sim$groups), K)