+++ tests/testthat/test-pagfl_results.R
@@ -51,10 +51,10 @@ test_that("pagfl PGMM output", {
   check_pagfl_output(estim = estim, X = X, i_index = data$i, t_index = data$t)
 })
 
-test_that("tv_pagfl results", {
+test_that("fuse_time results", {
   skip_on_cran()
   source(test_path("fixtures", "test_helper.R"))
-  sim <- readRDS(test_path("fixtures", "tv_pagfl_sim_2.rds"))
+  sim <- readRDS(test_path("fixtures", "fuse_time_sim_2.rds"))
   groups_0 <- sim$groups
   alpha_0 <- sim$alpha
   y <- sim$y
@@ -62,25 +62,25 @@ test_that("tv_pagfl results", {
   colnames(X) <- "a"
   p <- 2
   data <- as.data.frame(cbind(y = c(y), X))
-  estim <- tv_pagfl(y ~ 1 + a, data = data, n_periods = 100, lambda = 15, verbose = F)
-  check_tv_pagfl(estim = estim, groups_0 = groups_0, alpha_0 = alpha_0)
-  check_tv_pagfl_output(estim = estim, X = X)
+  estim <- fuse_time(y ~ 1 + a, data = data, n_periods = 100, lambda = 15, verbose = F)
+  check_fuse_time(estim = estim, groups_0 = groups_0, alpha_0 = alpha_0)
+  check_fuse_time_output(estim = estim, X = X)
   # With . formula
-  estim <- tv_pagfl(y ~ ., data = data, n_periods = 100, lambda = 15, verbose = F)
-  check_tv_pagfl(estim = estim, groups_0 = groups_0, alpha_0 = alpha_0)
-  check_tv_pagfl_output(estim = estim, X = X)
+  estim <- fuse_time(y ~ ., data = data, n_periods = 100, lambda = 15, verbose = F)
+  check_fuse_time(estim = estim, groups_0 = groups_0, alpha_0 = alpha_0)
+  check_fuse_time_output(estim = estim, X = X)
   # With index
   data$i <- rep(1:10, each = 100)
   data$t <- rep(1:100, 10)
-  estim <- tv_pagfl(y ~ ., data = data, index = c("i", "t"), lambda = 15, verbose = F)
-  check_tv_pagfl(estim = estim, groups_0 = groups_0, alpha_0 = alpha_0)
-  check_tv_pagfl_output(estim = estim, X = X, i_index = data$i, t_index = data$t)
+  estim <- fuse_time(y ~ ., data = data, index = c("i", "t"), lambda = 15, verbose = F)
+  check_fuse_time(estim = estim, groups_0 = groups_0, alpha_0 = alpha_0)
+  check_fuse_time_output(estim = estim, X = X, i_index = data$i, t_index = data$t)
 })
 
 
-test_that("tv_pagfl Unbalanced panel output", {
+test_that("fuse_time Unbalanced panel output", {
   skip_on_cran()
-  sim <- readRDS(test_path("fixtures", "tv_pagfl_sim.rds"))
+  sim <- readRDS(test_path("fixtures", "fuse_time_sim.rds"))
   y <- sim$y
   data <- as.data.frame(cbind(y = c(y)))
   data$i <- as.character(rep(1:20, each = 100))
@@ -93,21 +93,21 @@ test_that("tv_pagfl Unbalanced panel output", {
   data[data$t %in% c(1, 100), "y"] <- NA
   # Ensure that the second period for all but one group is omitted
   data[data$t == 2 & data$i != 1, ] <- NA
-  estim <- tv_pagfl(y ~ 1, data = data, index = c("i", "t"), lambda = 10, verbose = F)
+  estim <- fuse_time(y ~ 1, data = data, index = c("i", "t"), lambda = 10, verbose = F)
   # Time periods without support at the beginning or end are omitted from the output
   expect_equal(nrow(estim$coefficients$tv), 98)
   # Time periods without support at the beginning or end for some groups are given NA
   expect_equal(sum(is.na(estim$coefficients$tv[1, , ])), 2)
 })
 
-test_that("tv_pagfl const_coef", {
+test_that("fuse_time const_coef", {
   skip_on_cran()
-  sim <- readRDS(test_path("fixtures", "tv_pagfl_sim_2.rds"))
+  sim <- readRDS(test_path("fixtures", "fuse_time_sim_2.rds"))
   y <- sim$y
   X <- sim$X
   colnames(X) <- "a"
   data <- as.data.frame(cbind(y = c(y), X))
-  estim <- tv_pagfl(y ~ 1 + a, data = data, n_periods = 100, lambda = 20, const_coef = "a", verbose = F)
+  estim <- fuse_time(y ~ 1 + a, data = data, n_periods = 100, lambda = 20, const_coef = "a", verbose = F)
 
   expect_equal(dim(estim$coefficients$const), c(3, 1))
   expect_equal(colnames(estim$coefficients$const), "a")